+++ resources/views/admin/auth/index.php
@@ -10,6 +10,7 @@ declare(strict_types=1);
  */
 
 use Hirtz\Skeleton\Modules\Admin\Widgets\Grids\AuthItemGridView;
+use Hirtz\Skeleton\Modules\Admin\Widgets\HintAlert;
 use Hirtz\Skeleton\Web\View;
 use Hirtz\Skeleton\Widgets\Grids\GridContainer;
 use Hirtz\Skeleton\Widgets\Navs\Header;
@@ -17,6 +18,9 @@ use yii\data\ActiveDataProvider;
 
 echo Header::make()
     ->title(Yii::t('skeleton', 'COMMON_PERMISSIONS'));
+
+echo HintAlert::make()
+    ->text(Yii::t('skeleton', 'AUTH_INDEX_HINT'));
 
 echo GridContainer::make()
     ->grid(AuthItemGridView::make()